+++ cli/src/decode.rs
@@ -45,9 +45,9 @@ fn decode_voter(data: &[u8]) -> Result<()> {
         deposit_entries: voter
             .deposits
             .iter()
-            .filter(|d| d.is_used)
+            .filter(|d| d.is_used())
             .map(|d| DisplayDepositEntry {
-                allow_clawback: d.allow_clawback,
+                allow_clawback: d.allow_clawback(),
                 mint_index: d.voting_mint_config_idx,
                 unlocked_now: d.amount_unlocked(now_ts),
                 locked_now: d.amount_locked(now_ts),
@@ -66,7 +66,7 @@ fn decode_voter(data: &[u8]) -> Result<()> {
 // Read a sequence of base64 encoded accounts from stdin
 // and write their decoded versions back out as JSON.
 pub fn decode_account() -> Result<()> {
-    let account_types = HashMap::from([(Voter::discriminator(), &decode_voter)]);
+    let account_types = HashMap::from([(Voter::DISCRIMINATOR, &decode_voter)]);
 
     for line in io::stdin().lock().lines() {
         let data = base64::decode(line?)?;