+++ programs/voter-stake-registry/src/constants.rs
@@ -1,0 +1,7 @@
+use anchor_lang::solana_program::pubkey::Pubkey;
+use anchor_lang::pubkey;
+
+pub const ALLOWED_CPI_PROGRAMS: &[Pubkey] = &[
+    // Squads V4 Program
+    pubkey!("SQDS4ep65T869zMMBKyuUq6aD6EgTu8psMjkvj52pCf"),
+];