+++ programs/voter-stake-registry/src/instructions/create_voter.rs
@@ -1,3 +1,4 @@
+use crate::constants::ALLOWED_CPI_PROGRAMS;
 use crate::error::*;
 use crate::state::*;
 use anchor_lang::prelude::*;
@@ -56,23 +57,24 @@ pub fn create_voter(
 ) -> Result<()> {
     // Forbid creating voter accounts from CPI. The goal is to make automation
     // impossible that weakens some of the limitations intentionally imposed on
-    // locked tokens.
+    // locked tokens. Except from explicitly allowed programs.
     {
         let ixns = ctx.accounts.instructions.to_account_info();
         let current_index = tx_instructions::load_current_index_checked(&ixns)? as usize;
         let current_ixn = tx_instructions::load_instruction_at_checked(current_index, &ixns)?;
-        require_keys_eq!(
-            current_ixn.program_id,
-            *ctx.program_id,
-            VsrError::ForbiddenCpi
-        );
+
+        // Fast path: if the program_id matches this program, allow.
+        if current_ixn.program_id != *ctx.program_id {
+            // Otherwise, only allow if in ALLOWED_CPI_PROGRAMS.
+            require!(
+                ALLOWED_CPI_PROGRAMS.contains(&current_ixn.program_id),
+                VsrError::ForbiddenCpi
+            );
+        }
     }
 
-    require_eq!(voter_bump, *ctx.bumps.get("voter").unwrap());
-    require_eq!(
-        voter_weight_record_bump,
-        *ctx.bumps.get("voter_weight_record").unwrap()
-    );
+    require_eq!(voter_bump, ctx.bumps.voter);
+    require_eq!(voter_weight_record_bump, ctx.bumps.voter_weight_record);
 
     // Load accounts.
     let registrar = &ctx.accounts.registrar.load()?;
@@ -85,8 +87,7 @@ pub fn create_voter(
     voter.registrar = ctx.accounts.registrar.key();
 
     let voter_weight_record = &mut ctx.accounts.voter_weight_record;
-    voter_weight_record.account_discriminator =
-        spl_governance_addin_api::voter_weight::VoterWeightRecord::ACCOUNT_DISCRIMINATOR;
+    voter_weight_record.account_discriminator = VoterWeightRecord::discriminator();
     voter_weight_record.realm = registrar.realm;
     voter_weight_record.governing_token_mint = registrar.realm_governing_token_mint;
     voter_weight_record.governing_token_owner = voter_authority;