+++ programs/voter-stake-registry/src/instructions/set_time_offset.rs
@@ -12,6 +12,7 @@ pub struct SetTimeOffset<'info> {
 }
 
 /// A debug-only instruction that advances the time.
+#[allow(dead_code)]
 pub fn set_time_offset(ctx: Context<SetTimeOffset>, time_offset: i64) -> Result<()> {
     let allowed_program = Pubkey::from_str("GovernanceProgramTest1111111111111111111111").unwrap();
     let registrar = &mut ctx.accounts.registrar.load_mut()?;