+++ programs/voter-stake-registry/src/lib.rs
@@ -1,4 +1,4 @@
-#![allow(clippy::result_large_err)]
+#![allow(unexpected_cfgs)]
 
 use anchor_lang::prelude::*;
 use instructions::*;
@@ -9,12 +9,13 @@ pub mod events;
 mod governance;
 mod instructions;
 pub mod state;
+mod constants;
 
 #[macro_use]
 extern crate static_assertions;
 
 // The program address.
-declare_id!("4Q6WW2ouZ6V3iaNm56MTd5n2tnTm4C5fiH8miFHnAFHo");
+declare_id!("9SJqwCQ5AJkFtC7zxfFsF6Y5dm22XzN3JEhn3N14v23t");
 
 /// # Introduction
 ///
@@ -68,8 +69,8 @@ pub mod voter_stake_registry {
         instructions::create_registrar(ctx, registrar_bump)
     }
 
-    pub fn configure_voting_mint(
-        ctx: Context<ConfigureVotingMint>,
+    pub fn configure_voting_mint<'info>(
+        ctx: Context<'_, '_, 'info, 'info, ConfigureVotingMint<'info>>,
         idx: u16,
         digit_shift: i8,
         baseline_vote_weight_scaled_factor: u64,
@@ -197,13 +198,13 @@ pub mod voter_stake_registry {
         instructions::update_voter_weight_record(ctx)
     }
 
-    pub fn update_max_vote_weight(ctx: Context<UpdateMaxVoteWeight>) -> Result<()> {
+    pub fn update_max_vote_weight<'info>(
+        ctx: Context<'_, '_, 'info, 'info, UpdateMaxVoteWeight<'info>>,
+    ) -> Result<()> {
         instructions::update_max_vote_weight(ctx)
     }
 
-    pub fn close_voter<'key, 'accounts, 'remaining, 'info>(
-        ctx: Context<'key, 'accounts, 'remaining, 'info, CloseVoter<'info>>,
-    ) -> Result<()> {
+    pub fn close_voter<'info>(ctx: Context<'_, '_, 'info, 'info, CloseVoter<'info>>) -> Result<()> {
         instructions::close_voter(ctx)
     }
 
@@ -215,6 +216,8 @@ pub mod voter_stake_registry {
         instructions::log_voter_info(ctx, deposit_entry_begin, deposit_entry_count)
     }
 
+    // this is a temporary function to set the time offset for testing
+    // it should be removed in the prod version
     pub fn set_time_offset(ctx: Context<SetTimeOffset>, time_offset: i64) -> Result<()> {
         instructions::set_time_offset(ctx, time_offset)
     }