+++ programs/voter-stake-registry/tests/program_test/mod.rs
@@ -2,18 +2,19 @@ use log::*;
 use std::cell::RefCell;
 use std::{str::FromStr, sync::Arc, sync::RwLock};
 
-use solana_program::{program_option::COption, program_pack::Pack};
+use anchor_lang::solana_program::{program_option::COption, program_pack::Pack};
 use solana_program_test::*;
 use solana_sdk::{
     pubkey::Pubkey,
     signature::{Keypair, Signer},
 };
-use spl_token::{state::*, *};
+use spl_token::state::*;
 
 pub use addin::*;
 pub use cookies::*;
 pub use governance::*;
 pub use solana::*;
+#[allow(unused_imports)]
 pub use utils::*;
 
 pub mod addin;
@@ -79,6 +80,7 @@ impl Log for LoggerWrapper {
     fn flush(&self) {}
 }
 
+#[allow(dead_code)]
 pub struct TestContext {
     pub solana: Arc<SolanaCookie>,
     pub governance: GovernanceCookie,
@@ -107,11 +109,7 @@ impl TestContext {
 
         let addin_program_id = voter_stake_registry::id();
 
-        let mut test = ProgramTest::new(
-            "voter_stake_registry",
-            addin_program_id,
-            processor!(voter_stake_registry::entry),
-        );
+        let mut test = ProgramTest::new("voter_stake_registry", addin_program_id, None);
         // intentionally set to half the limit, to catch potential problems early
         test.set_compute_max_units(120000);
 
@@ -120,7 +118,7 @@ impl TestContext {
         test.add_program(
             "spl_governance",
             governance_program_id,
-            processor!(spl_governance::processor::process_instruction),
+            None,
         );
 
         // Setup the environment
@@ -210,7 +208,7 @@ impl TestContext {
             });
         }
 
-        let mut context = test.start_with_context().await;
+        let context = test.start_with_context().await;
         let rent = context.banks_client.get_rent().await.unwrap();
 
         let solana = Arc::new(SolanaCookie {