+++ programs/voter-stake-registry/tests/test_deposit_monthly_vesting.rs
@@ -40,7 +40,7 @@ async fn balances(
     }
 }
 
-#[allow(unaligned_references)]
+
 #[tokio::test]
 async fn test_deposit_monthly_vesting() -> Result<(), TransportError> {
     let context = TestContext::new().await;