+++ programs/voter-stake-registry/tests/test_log_voter_info.rs
@@ -8,13 +8,13 @@ mod program_test;
 
 fn deserialize_event<T: anchor_lang::Event>(event: &str) -> Option<T> {
     let data = base64::decode(event).ok()?;
-    if data.len() < 8 || data[0..8] != T::discriminator() {
+    if data.len() < 8 || &data[0..8] != T::DISCRIMINATOR {
         return None;
     }
     T::try_from_slice(&data[8..]).ok()
 }
 
-#[allow(unaligned_references)]
+
 #[tokio::test]
 async fn test_print_event() -> Result<(), TransportError> {
     println!(
@@ -27,7 +27,7 @@ async fn test_print_event() -> Result<(), TransportError> {
     Ok(())
 }
 
-#[allow(unaligned_references)]
+
 #[tokio::test]
 async fn test_log_voter_info() -> Result<(), TransportError> {
     let context = TestContext::new().await;