+++ programs/voter-stake-registry/tests/test_voting.rs
@@ -1,4 +1,6 @@
-use anchor_spl::token::TokenAccount;
+use anchor_spl::{
+    token::TokenAccount, token_interface::spl_token_metadata_interface::borsh::BorshDeserialize,
+};
 use program_test::*;
 use solana_program_test::*;
 use solana_sdk::{signature::Keypair, signer::Signer, transport::TransportError};
@@ -6,7 +8,6 @@ use voter_stake_registry::state::LockupKind;
 
 mod program_test;
 
-#[allow(unaligned_references)]
 #[tokio::test]
 async fn test_voting() -> Result<(), TransportError> {
     let context = TestContext::new().await;
@@ -236,7 +237,7 @@ async fn test_voting() -> Result<(), TransportError> {
     let proposal_data = context.solana.get_account_data(proposal.address).await;
     let mut data_slice: &[u8] = &proposal_data;
     let proposal_state: spl_governance::state::proposal::ProposalV2 =
-        anchor_lang::AnchorDeserialize::deserialize(&mut data_slice).unwrap();
+        BorshDeserialize::deserialize(&mut data_slice).unwrap();
     assert_eq!(proposal_state.options[0].vote_weight, 2 * 750);
     assert_eq!(proposal_state.deny_vote_weight.unwrap(), 0);
 
@@ -276,6 +277,7 @@ async fn test_voting() -> Result<(), TransportError> {
             voter2.token_owner_record,
             &voter2_authority,
             payer.pubkey(),
+            realm.clone(),
         )
         .await
         .unwrap();